Agrega template Tags para determinar si es o no subscriptor o administrador
