--- cmsis2/templatetags/has_group.py
@@ -1,10 +1,0 @@
-from django import template
-from django.contrib.auth.models import Group
-
-register = template.Library()
-
-
-@register.filter(name='has_group')
-def has_group(user, group_name):
-    group = Group.objects.get(name=group_name)
-    return True if group in user.groups.all() else False
